requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
autopep8==2.0.4
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.1
django-silk==5.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docopt==0.6.2
drf-nested-routers==0.93.5
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
flower==2.0.1
gevent==24.2.1
geventhttpclient==2.0.11
gprof2dot==2022.7.29
greenlet==3.0.3
gunicorn==21.2.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
locust==2.24.0
MarkupSafe==2.1.5
model-bakery==1.17.0
msgpack==1.0.8
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
pyzmq==25.1.2
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
roundrobin==0.0.4
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.2
